refact: #191 다른 탭에도 사용 가능하도록 공통 훅으로 분리
